+++ packages/stream-tile-handler/src/__tests__/three-id.test.ts
@@ -37,7 +37,7 @@ const FAKE_CID_2 = new CID('bafybeig6xv5nwphfmvcnektpnojts44jqcuam7bmye2pb54adnr
 const FAKE_CID_3 = new CID('bafybeig6xv5nwphfmvcnektpnojts55jqcuam7bmye2pb54adnrtccjlsu')
 const FAKE_CID_4 = new CID('bafybeig6xv5nwphfmvcnektpnojts66jqcuam7bmye2pb54adnrtccjlsu')
 
-const RECORDS = {
+const COMMITS = {
   genesis: {
     header: {
       tags: ['3id'],
@@ -71,7 +71,7 @@ const RECORDS = {
   },
   r1: {
     desiredContent: { publicKeys: { test: '0xabc' }, other: 'data' },
-    record: {
+    commit: {
       jws: {
         payload: 'bbbb',
         signatures: [
@@ -97,7 +97,7 @@ const RECORDS = {
       },
     },
   },
-  r2: { record: { proof: FAKE_CID_4 } },
+  r2: { commit: { proof: FAKE_CID_4 } },
   proof: {
     blockNumber: 123456,
     blockTimestamp: 1615799679,
@@ -211,17 +211,17 @@ it('is constructed correctly', async () => {
   expect(tileDocumentHandler.name).toEqual('tile')
 })
 
-it('makes genesis record correctly', async () => {
-  const record = (await TileDocument.makeGenesis(context.api, RECORDS.genesis.data, {
-    ...RECORDS.genesis.header,
+it('makes genesis commit correctly', async () => {
+  const commit = (await TileDocument.makeGenesis(context.api, COMMITS.genesis.data, {
+    ...COMMITS.genesis.header,
   })) as SignedCommitContainer
-  const { jws, linkedBlock } = record
+  const { jws, linkedBlock } = commit
 
   const payload = dagCBOR.util.deserialize(linkedBlock)
   const generated = { jws: serialize(jws), linkedBlock: serialize(payload) }
 
-  // Add the 'unique' header field to the data used to generate the expected genesis record
-  const genesis = cloneDeep(RECORDS.genesis)
+  // Add the 'unique' header field to the data used to generate the expected genesis commit
+  const genesis = cloneDeep(COMMITS.genesis)
   genesis.header['unique'] = generated.linkedBlock.header.unique
 
   const expected = await did.createDagJWS(genesis)
@@ -234,34 +234,34 @@ it('makes genesis record correctly', async () => {
   expect(generated).toEqual(serialize(signed))
 })
 
-it('applies genesis record correctly', async () => {
+it('applies genesis commit correctly', async () => {
   const tileHandler = new TileDocumentHandler()
 
-  const record = (await TileDocument.makeGenesis(context.api, RECORDS.genesis.data, {
+  const commit = (await TileDocument.makeGenesis(context.api, COMMITS.genesis.data, {
     controllers: [did.id],
     tags: ['3id'],
   })) as SignedCommitContainer
-  await context.ipfs.dag.put(record, FAKE_CID_1)
+  await context.ipfs.dag.put(commit, FAKE_CID_1)
 
-  const payload = dagCBOR.util.deserialize(record.linkedBlock)
-  await context.ipfs.dag.put(payload, record.jws.link)
+  const payload = dagCBOR.util.deserialize(commit.linkedBlock)
+  await context.ipfs.dag.put(payload, commit.jws.link)
 
-  const signedCommitData = { cid: FAKE_CID_1, type: CommitType.GENESIS, commit: payload, envelope: record.jws }
+  const signedCommitData = { cid: FAKE_CID_1, type: CommitType.GENESIS, commit: payload, envelope: commit.jws }
   const streamState = await tileHandler.applyCommit(signedCommitData, context)
   delete streamState.metadata.unique
   expect(streamState).toMatchSnapshot()
 })
 
-it('makes signed record correctly', async () => {
+it('makes signed commit correctly', async () => {
   const tileDocumentHandler = new TileDocumentHandler()
 
-  await context.ipfs.dag.put(RECORDS.genesisGenerated.jws, FAKE_CID_1)
+  await context.ipfs.dag.put(COMMITS.genesisGenerated.jws, FAKE_CID_1)
   await context.ipfs.dag.put(
-    RECORDS.genesisGenerated.linkedBlock,
-    RECORDS.genesisGenerated.jws.link
+    COMMITS.genesisGenerated.linkedBlock,
+    COMMITS.genesisGenerated.jws.link
   )
 
-  const genesisCommitData = { cid: FAKE_CID_1, type: CommitType.GENESIS, commit: RECORDS.genesisGenerated.linkedBlock, envelope: RECORDS.genesisGenerated.jws }
+  const genesisCommitData = { cid: FAKE_CID_1, type: CommitType.GENESIS, commit: COMMITS.genesisGenerated.linkedBlock, envelope: COMMITS.genesisGenerated.jws }
   const state = await tileDocumentHandler.applyCommit(
     genesisCommitData,
     context
@@ -269,49 +269,49 @@ it('makes signed record correctly', async () => {
   const state$ = TestUtils.runningState(state)
   const doc = new TileDocument(state$, context)
 
-  await expect(doc.makeCommit({} as CeramicApi, RECORDS.r1.desiredContent)).rejects.toThrow(
+  await expect(doc.makeCommit({} as CeramicApi, COMMITS.r1.desiredContent)).rejects.toThrow(
     /No DID/
   )
 
-  const record = (await doc.makeCommit(
+  const commit = (await doc.makeCommit(
     context.api,
-    RECORDS.r1.desiredContent
+    COMMITS.r1.desiredContent
   )) as SignedCommitContainer
-  const { jws: rJws, linkedBlock: rLinkedBlock } = record
+  const { jws: rJws, linkedBlock: rLinkedBlock } = commit
   const rPayload = dagCBOR.util.deserialize(rLinkedBlock)
-  expect({ jws: serialize(rJws), payload: serialize(rPayload) }).toEqual(RECORDS.r1.record)
+  expect({ jws: serialize(rJws), payload: serialize(rPayload) }).toEqual(COMMITS.r1.commit)
 })
 
-it('applies signed record correctly', async () => {
+it('applies signed commit correctly', async () => {
   const tileDocumentHandler = new TileDocumentHandler()
 
-  const genesisRecord = (await TileDocument.makeGenesis(context.api, RECORDS.genesis.data, {
+  const genesisCommit = (await TileDocument.makeGenesis(context.api, COMMITS.genesis.data, {
     controllers: [did.id],
     tags: ['3id'],
   })) as SignedCommitContainer
-  await context.ipfs.dag.put(genesisRecord, FAKE_CID_1)
+  await context.ipfs.dag.put(genesisCommit, FAKE_CID_1)
 
-  const payload = dagCBOR.util.deserialize(genesisRecord.linkedBlock)
-  await context.ipfs.dag.put(payload, genesisRecord.jws.link)
+  const payload = dagCBOR.util.deserialize(genesisCommit.linkedBlock)
+  await context.ipfs.dag.put(payload, genesisCommit.jws.link)
 
   // apply genesis
-  const genesisCommitData = { cid: FAKE_CID_1, type: CommitType.GENESIS, commit: payload, envelope: genesisRecord.jws }
+  const genesisCommitData = { cid: FAKE_CID_1, type: CommitType.GENESIS, commit: payload, envelope: genesisCommit.jws }
   let state = await tileDocumentHandler.applyCommit(genesisCommitData, context)
 
   const state$ = TestUtils.runningState(state)
   const doc = new TileDocument(state$, context)
-  const signedRecord = (await doc.makeCommit(
+  const signedCommit = (await doc.makeCommit(
     context.api,
-    RECORDS.r1.desiredContent
+    COMMITS.r1.desiredContent
   )) as SignedCommitContainer
 
-  await context.ipfs.dag.put(signedRecord, FAKE_CID_2)
+  await context.ipfs.dag.put(signedCommit, FAKE_CID_2)
 
-  const sPayload = dagCBOR.util.deserialize(signedRecord.linkedBlock)
-  await context.ipfs.dag.put(sPayload, signedRecord.jws.link)
+  const sPayload = dagCBOR.util.deserialize(signedCommit.linkedBlock)
+  await context.ipfs.dag.put(sPayload, signedCommit.jws.link)
 
   // apply signed
-  const signedCommitData = { cid: FAKE_CID_2, type: CommitType.SIGNED, commit: sPayload, envelope: signedRecord.jws }
+  const signedCommitData = { cid: FAKE_CID_2, type: CommitType.SIGNED, commit: sPayload, envelope: signedCommit.jws }
   state = await tileDocumentHandler.applyCommit(
     signedCommitData,
     context,
@@ -322,63 +322,63 @@ it('applies signed record correctly', async () => {
   expect(state).toMatchSnapshot()
 })
 
-it('throws error if record signed by wrong DID', async () => {
+it('throws error if commit signed by wrong DID', async () => {
   const tileDocumentHandler = new TileDocumentHandler()
 
-  const genesisRecord = (await TileDocument.makeGenesis(context.api, RECORDS.genesis.data, {
+  const genesisCommit = (await TileDocument.makeGenesis(context.api, COMMITS.genesis.data, {
     controllers: ['did:3:fake'],
     tags: ['3id'],
   })) as SignedCommitContainer
-  await context.ipfs.dag.put(genesisRecord, FAKE_CID_1)
+  await context.ipfs.dag.put(genesisCommit, FAKE_CID_1)
 
-  const payload = dagCBOR.util.deserialize(genesisRecord.linkedBlock)
-  await context.ipfs.dag.put(payload, genesisRecord.jws.link)
+  const payload = dagCBOR.util.deserialize(genesisCommit.linkedBlock)
+  await context.ipfs.dag.put(payload, genesisCommit.jws.link)
 
-  const genesisCommitData = { cid: FAKE_CID_1, type: CommitType.GENESIS, commit: payload, envelope: genesisRecord.jws }
+  const genesisCommitData = { cid: FAKE_CID_1, type: CommitType.GENESIS, commit: payload, envelope: genesisCommit.jws }
   await expect(
     tileDocumentHandler.applyCommit(genesisCommitData, context)
   ).rejects.toThrow(/invalid_jws: not a valid verificationMethod for issuer/)
 })
 
-it('applies anchor record correctly', async () => {
+it('applies anchor commit correctly', async () => {
   const tileDocumentHandler = new TileDocumentHandler()
 
-  const genesisRecord = (await TileDocument.makeGenesis(context.api, RECORDS.genesis.data, {
+  const genesisCommit = (await TileDocument.makeGenesis(context.api, COMMITS.genesis.data, {
     controllers: [did.id],
     tags: ['3id'],
   })) as SignedCommitContainer
-  await context.ipfs.dag.put(genesisRecord, FAKE_CID_1)
+  await context.ipfs.dag.put(genesisCommit, FAKE_CID_1)
 
-  const payload = dagCBOR.util.deserialize(genesisRecord.linkedBlock)
-  await context.ipfs.dag.put(payload, genesisRecord.jws.link)
+  const payload = dagCBOR.util.deserialize(genesisCommit.linkedBlock)
+  await context.ipfs.dag.put(payload, genesisCommit.jws.link)
 
   // apply genesis
-  const genesisCommitData = { cid: FAKE_CID_1, type: CommitType.GENESIS, commit: payload, envelope: genesisRecord.jws }
+  const genesisCommitData = { cid: FAKE_CID_1, type: CommitType.GENESIS, commit: payload, envelope: genesisCommit.jws }
   let state = await tileDocumentHandler.applyCommit(genesisCommitData, context)
 
   const state$ = TestUtils.runningState(state)
   const doc = new TileDocument(state$, context)
-  const signedRecord = (await doc.makeCommit(
+  const signedCommit = (await doc.makeCommit(
     context.api,
-    RECORDS.r1.desiredContent
+    COMMITS.r1.desiredContent
   )) as SignedCommitContainer
 
-  await context.ipfs.dag.put(signedRecord, FAKE_CID_2)
+  await context.ipfs.dag.put(signedCommit, FAKE_CID_2)
 
-  const sPayload = dagCBOR.util.deserialize(signedRecord.linkedBlock)
-  await context.ipfs.dag.put(sPayload, signedRecord.jws.link)
+  const sPayload = dagCBOR.util.deserialize(signedCommit.linkedBlock)
+  await context.ipfs.dag.put(sPayload, signedCommit.jws.link)
 
   // apply signed
-  const signedCommitData = { cid: FAKE_CID_2, type: CommitType.SIGNED, commit: sPayload, envelope: signedRecord.jws }
+  const signedCommitData = { cid: FAKE_CID_2, type: CommitType.SIGNED, commit: sPayload, envelope: signedCommit.jws }
   state = await tileDocumentHandler.applyCommit(
     signedCommitData,
     context,
     state
   )
 
-  await context.ipfs.dag.put(RECORDS.proof, FAKE_CID_4)
+  await context.ipfs.dag.put(COMMITS.proof, FAKE_CID_4)
   // apply anchor
-  const anchorCommitData = { cid: FAKE_CID_3, type: CommitType.ANCHOR, commit: RECORDS.r2.record, proof: RECORDS.proof }
+  const anchorCommitData = { cid: FAKE_CID_3, type: CommitType.ANCHOR, commit: COMMITS.r2.commit, proof: COMMITS.proof }
   state = await tileDocumentHandler.applyCommit(
     anchorCommitData,
     context,