+++ packages/blockchain-utils-linking/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/blockchain-utils-linking@1.1.4-rc.0...@ceramicnetwork/blockchain-utils-linking@1.2.0-rc.0) (2021-09-13)
+
+
+### Features
+
+* caip10 ethereum+eos integration test ([#1668](https://github.com/ceramicnetwork/js-ceramic/issues/1668)) ([4267d8b](https://github.com/ceramicnetwork/js-ceramic/commit/4267d8b31fca38a163bb009198a49e1de22b2a58)), closes [#1678](https://github.com/ceramicnetwork/js-ceramic/issues/1678)
+
+
+
+
+
 ## [1.1.4-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/blockchain-utils-linking@1.1.3...@ceramicnetwork/blockchain-utils-linking@1.1.4-rc.0) (2021-09-08)
 
 **Note:** Version bump only for package @ceramicnetwork/blockchain-utils-linking