+++ packages/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0-rc.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/core@1.6.0-rc.0...@ceramicnetwork/core@1.6.0-rc.1) (2021-09-13)
+
+**Note:** Version bump only for package @ceramicnetwork/core
+
+
+
+
+
 # [1.6.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/core@1.5.0...@ceramicnetwork/core@1.6.0-rc.0) (2021-09-08)