+++ packages/stream-caip10-link/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.4-rc.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-caip10-link@1.1.4-rc.0...@ceramicnetwork/stream-caip10-link@1.1.4-rc.1) (2021-09-13)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-caip10-link
+
+
+
+
+
 ## [1.1.4-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-caip10-link@1.1.3...@ceramicnetwork/stream-caip10-link@1.1.4-rc.0) (2021-09-08)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-caip10-link